Make integration test router bind to random port
resolves #4720

Previously, the integration test would spin up the router with its
listener always binding to port 4000, which might cause tests to
fail if a different process were to bind to the same port.

This commit makes the test router always bind to an available port
and corrects all calls to the router to use this new port.