Fix chunk formatting in multipart protocol (#4681)
This PR changes the way we're sending chunks in the stream. Instead of
finishing the chunk with `\r\n` we don't send this at the end of our
current chunk but instead at the beginning of the next one. For the end
users nothing changes but it let us to close the stream with the right
final boundary by appending `--\r\n` directly to the last chunk.

Fixes #4634 

---------

Signed-off-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>