refresh query plan expiration on GET (#4604)
Fix https://github.com/apollographql/router/issues/4473

We want the query plan cache to act like a LRU, so if a TTL is set in
its Redis configuration, we will reset the plan cache key's current
expiration when reading it.

This is now part of the common redis configuration and is enabled by default. For entity caching, this option is
disabled, as the cache will manage TTL directly

Integration tests now require Redis 7. Otherwise you will get the error "ERR unknown command `EXPIRETIME`"