drop the ExtensionsGuard earlier (#4805)
Now that UsageReporting is stored in an Arc, we can just clone it and
drop the lock, then insert what we want in the context outside of the
lock

This fixes a bug often triggerd in CI: 

```
thread 'test_plugin_ordering' panicked at apollo-router/src/context/extensions/sync.rs:66:21:
ExtensionsGuard held for 11ms. This is probably a bug that will stall the Router and cause performance problems. Run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```