add batching cancellations to coprocessors
This commit allows cancelled requests by a coprocessor to also
cancel portions of a batch query correctly.

This logic should eventually be moved higher up the call stack of
the router so that future plugins, custom plugins, and anything
else that might cancel portions of a batch query not timeout
without needing to copy this commit's logic.