Implement basic static cost calculator for demand control (#4829)
# Basic Static Cost Estimation

Implements a basic cost calculator for demand control. Given a
supergraph schema and a query, the query document is recursively scored
with the mapping:

Operations:
- Mutation: 10
- Query: 0
- Subscription: 0

Fields:
- Object: 1
- Interface: 1
- Union: 1
- Scalar: 0
- Enum: 0

## Handling Directives

While traversing the document, the calculator accounts for uses of @skip
and @include which exclude fields from the response, and thus the cost.

Additionally, the federations @requires directive adds the cost of the selection
being required to the cost of the field that requires it.  This implementation
does not account for double counting. So two fields requiring the same
selection will result in that requirement contributing twice to the total cost.

## Handling Lists

Without user annotations about paging limitations, the static analysis
cannot determine the length of list fields. Until we define paging
directives to specify this, we assume a constant upper bound of 100 for
all list fields. Future updates will include using a new directive to
indicate the list size approximation as well as calculating a dynamic
cost from the response data returned by the subgraphs.

Fixes
[ROUTER-171](https://apollographql.atlassian.net/browse/ROUTER-171)

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [X] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [X] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.


[ROUTER-171]:
https://apollographql.atlassian.net/browse/ROUTER-171?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ