Add `experimental_log_broken_pipe` (#4870)
## Experimental logging of broken pipe errors

You can emit a log message each time the client closes the connection
early, which can help you debug issues with clients that close
connections before the server can respond.

This feature is disabled by default but can be enabled by setting the
`experimental_log_broken_pipe` option to `true`:

```yaml title="router.yaml"
supergraph:
  experimental_log_broken_pipe: true
```
| Attribute | Default | Description |

|----------------------------------|---------|-----------------------------------------------------|
| `experimental_log_broken_pipe` | false | Emit a log message if a
broken pipe was detected. |



#4770 introduced a log message when a broken pipe is detected. This PR
makes this log message opt-in.

Users that have internet facing routers will likely not want to opt in
to this log message as they have no control over the clients.


<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: bryn <bryn@apollographql.com>