fix annoying clippy warning on macOS
I don't think it's a good idea to `cfg` out integration tests by
platform. This is a good example of why not.

If we don't run tests to run on certain platforms, then we should check
at runtime to prevent irritating warnings/errors like this from creeping
into the codebase.