impl Display for OpGraphPath (https://github.com/apollographql/federation-next/pull/243)
This PR addresses Jira ticket FED-20.

Display for `OpGraphPath` has been implemented by representing the path
with an array of edges. The edges are represented with their Debug impl,
which can be changed. Currently, this implementation is the same as the
debug impl of the edges.

Since the `edges` are `Option<EdgeIndex>>`, I am unsure how to represent
the `None` case or if that was a concern at all. If there is a preferred
way to represent the edges, I'll gladly change it.