Demand Control: Calculate actual cost for a response (#4903)
Introduces a third method to `CostCalculator` which scores the cost of a
GraphQL response.

The first stage of scoring a response is to zip it together with the
request. The initial implementation does not use the request fields, but
this will be required to support custom cost. In that case, we will need
to check the corresponding request field definition to see if a custom
cost needs to be applied to a particular response field. This
information is stored in a counterpart to `serde_json::Value`, which is
called `TypedValue`. This `TypedValue` enum pairs each JSON value with
the corresponding `apollo_compiler::executable::Field`.

Once the response values are paired with their corresponding fields, it
is simple to traverse the JSON structure, taking query or schema
directives into account.

Implements
[ROUTER-175](https://apollographql.atlassian.net/browse/ROUTER-175)

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [X] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [X] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.


[ROUTER-175]:
https://apollographql.atlassian.net/browse/ROUTER-175?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ