transient in memory cache
This adds another level of in memory cache, to mitigate issues with
unique or infrequent queries pushing frequently used queries out of the
in memory cache.
If a query has only been seen once, then it is stored inthe transient
cache (short term, small number of entries, LRU).
If that query is requested again, and is still present in the transient
cache, then it is added to the long term, larger in memory cache, to the
redis cache as well, and removed from the transient cache.
If that query is requested again but not present in the transient cache,
then we test the larger in memory cache and redis