Enable Rust-based API schema generation by default (#4931)
Fixes #4649

Requires an apollo-federation update to feign support for join spec
v0.4. The new join spec doesn't affect how the API schema is generated
so the only thing the update does is *not error*. The new release also
includes a query plan Display implementation for #4815!

I ran it through our whole harness testing again, which is quite
reliable for API schema as the schema data is complete. The only
observable differences are in the sorting of keys in input objects in
default values which is both very limited, and doesn't make sense in the
JS version, so that's good. The *meaning* of schemas did not change in
any case.