Remove legacy validation (#4551)
Fix #4407 
Fix #4409
This removes GraphQL validation from the query planner, to use the Rust
version instead. Validation has now moved to the query analysis layer,
which means we can remove a lot of code that was there to accommodate
parsing and validating the query but not rejecting it outright, because
we needed to compare the validation result with the planner's.
This will greatly reduce the load on the planner, which will now only be
used for planning queries, not validating.
This new validation process has been running in production for months
concurrently with the JavaScript version, allowing us to detect and fix
any discrepancies in the new implementation. We now have enough
confidence in the new Rust-based validation to entirely switch off the
less performant, JavaScript validation.

Remaining issues:

- [x] usage reporting: since invalid queries are rejected in query
analysis, inside the router service, then usage reporting for invalid
queries is no longer reported, because the telemetry plugins handles
usage reporting at the supergraph response level
- [x] persisted queries: apparently persisted queries use the query
analysis transformation of supergraph requests, now one test fails on an
invalid query, I have not yet looked at why
- [x] `@defer` directive: a test is failing because the directive is
used in the query but not found by the schema. @lrlna what is
apollo-compiler expecting here? Is `@defer` considered builtin, or does
it need to be declared in the schema?

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: Iryna Shestak <shestak.irina@gmail.com>
Co-authored-by: Bryn Cooke <BrynCooke@gmail.com>
Co-authored-by: Simon Sapin <simon@apollographql.com>
Co-authored-by: Renée Kooi <renee.kooi@apollographql.com>