Move PluginTestHarness out of telemetry for use in other plugins (#4932)
This is suited for unit testing where you wish to inspect the request
and response both at the start of the pipeline and at the destination
service.

The existing TestHarness is about firing up the entire router stack.
Alternatives that we have tried are mocking, which never ends well.

This test harness is super simple, and allows you to create a plugin,
fire a fake request through, perform assertion on the request during the
pipeline execution, return a fake response, perform validation on the
returned response.

It's only used in telemetry at the moment, but it'll be useful for
demand control.

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

Co-authored-by: bryn <bryn@apollographql.com>