[batching] make batching look less racy (#4941)
It turns out that it's tricky. We need to use tokio async mutex to stop
the compiler complaining that parking lot mutex guard isn't send.

We still need the parking lot mutex, because we can't make
query_for_index() async or that causes a different Send/Sync issue.

Other problems are avoided by using:
 - AtomicUsize to track remaining
 - Arc/Mutex on Sender to allow for cloning and interior mutability

I'm not 100% convinced this is an improvement, since the code is slower
and more complex and I don't actually think there is a race in the
existing code *unless* someone is doing something very funky with the
batching extensions(). So, I'll put it on a branch and seek specific
feedback on this change.