allow query deduplication and batching to work together
If query deduplication was enabled, it would interfere with query
batching execution and cause failure with a timeout.

The two features now work together with batching taking precedence. Thus
if an individual query is part of a batch, deduplication (if enabled) is
by-passed (since it doesn't make sense in a batching context).

If both are enabled:
 - query deduplication will apply to all non-batched requests
 - batching will apply to all batched requests