Mimick graphql-js validation errors in new validation implementation
This is a best-effort change to match the messages graphql-js produces,
so users of the router who are relying on the message contents are less
likely to have to change their tests / snapshots / code.

In the future we will add stable error codes for validation errors, and
strongly discourage relying on the contents of error messages.