Rust validation follow-up: backwards compatibility fixes (#4949)
Error messages are now mostly the same as the graphql-js messages from
before (where possible). GraphQL syntax errors are reported as
`PARSING_ERROR` again instead of as `GRAPHQL_VALIDATION_ERROR`.

`SpecError::ParsingError` is replaced by `SpecError::ParseError` which
contains structured error information similar to
`SpecError::ValidationError`. The other uses of `ParsingError` are
replaced by `TransformError`, as most of them are in query
transforms/filtering. The error messages for `TransformError` remain the
same.

There will still be differences in error messages, especially when
querying undefined fields etc, because previously those would hit an
error case deep inside the router whereas now they will be rejected
early on. This is an easily explainable difference with obvious benefits
for users, so I think it's acceptable.

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.