file-uploads: Add tests and improve errors for rearrange_query_plan (#4920)
1. Replace unordered HashSet/HashMap with IndexSet/IndexMap. It makes
predictable order of variable names inside error messages
2. Add unit tests
3. For Parallel node simplify result query plan by doing special
handling of cases: a. if only one subnode is using files don't change
anything b. if all subnodes are using files replace `Parallel` with
`Sequence`

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.