Add support for condition processing to subgraph batching
In consultation with federation team, I had been under the impression
that we didn't need support for ConditionNode in the batching
implementation.

It turns out that in certain use cases, ConditionNode will arrive in a
batch and we need to decide how to deal with it.

This is a partially implemented solution to start discussion.