Parse fetch node operations into executable docs during planning (#4975)
In query plan fetch nodes, store a lazily-initialized
`ExecutableDocument` in addition to its GraphQL string representation.
Both the authorization plugin and demand control plugin were parsing
operations for each request. This `ExecutableDocument` is skipped in
serde (de)serialization, either when reading a JSON query plan from the
JS planner or when using the Redis distributed cache.

Fixes
[ROUTER-221](https://apollographql.atlassian.net/browse/ROUTER-221)

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.


[ROUTER-221]:
https://apollographql.atlassian.net/browse/ROUTER-221?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Simon Sapin <simon@apollographql.com>